+++ clinicadl_zoo/transforms/transforms/nan_removal.py
@@ -1,4 +1,4 @@
-from typing import Optional, get_type_hints
+from typing import Optional
 
 import torch
 import torchio as tio
