+++ tests/unittests/transforms/factory/test_nan_removal.py
@@ -5,7 +5,6 @@ from clinicadl.data.structures import DataPoint
 
 from clinicadl_zoo.transforms import NanRemoval
 from clinicadl_zoo.transforms.config import NanRemovalConfig
-from clinicadl_zoo.transforms.factory import get_transform_config
 
 GOOD_INPUTS = [
     {"posinf": 1.2, "neginf": 0.1},
@@ -55,11 +54,3 @@ def test_nan_removal():
     assert transformed.image.tensor[0, 0, 0, 0] == 1
     assert transformed.image.tensor[0, 1, 1, 0] == 1
     assert transformed.image.tensor[0, 1, 0, 0] == -1
-
-
-def test_factory():
-    config = get_transform_config("NanRemoval", nan=1)
-    assert config.name == "NanRemoval"
-    assert config.nan == 1
-    assert config.posinf is None
-    assert config.neginf is None
